% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bitly_auth.R
\name{bitly_get_token}
\alias{bitly_get_token}
\title{Retrieve the active Bitly token}
\usage{
bitly_get_token()
}
\value{
An invisible character string containing the active Bitly token.
}
\description{
Retrieves the Bitly personal access token used for authentication with the
Bitly API.

The token is searched for in the following order:
\enumerate{
  \item The \code{BITLY_PAT} environment variable.
  \item A local file at \code{~/.bitly/token}.
}

If a token file is found, it is loaded automatically using
\code{bitly_set_token()}.
}
\details{
If no token can be located, an error is raised instructing the user to set
a token manually.
}
\examples{
\dontrun{
bitly_get_token()
}

}
\seealso{
\code{\link{bitly_set_token}},
\code{\link{bitly_reset_token}},
\code{\link{bitly_test_token}}
}
