% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bitly_auth.R
\name{bitly_reset_token}
\alias{bitly_reset_token}
\title{Remove the active Bitly token}
\usage{
bitly_reset_token()
}
\value{
Invisibly returns \code{NULL}. Called for side effects.
}
\description{
Removes the \code{BITLY_PAT} environment variable from the current R session,
effectively clearing the active Bitly token.
}
\examples{
\dontrun{
bitly_reset_token()
}

}
\seealso{
\code{\link{bitly_set_token}},
\code{\link{bitly_get_token}}
}
