% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bitly_api.R
\name{bitly_shorten}
\alias{bitly_shorten}
\title{Create a Bitly short link}
\usage{
bitly_shorten(url, guid = bitly_get_groups()[1])
}
\arguments{
\item{url}{Character string giving the URL to shorten.}

\item{guid}{Character string giving the Bitly group GUID under which the
short link should be created. Defaults to the first available group.}
}
\value{
A character string containing the shortened Bitly URL.
}
\description{
Creates a shortened Bitly link for a supplied URL using the authenticated
Bitly account.

By default, the short link is created using the first available Bitly group
returned by \code{bitly_get_groups()}.
}
\details{
A success message is displayed when the short link is created.

This function requires a valid Bitly personal access token configured via
\code{bitly_get_token()}.
}
\examples{
\dontrun{
# Shorten a local livecode URL
bitly_shorten("http://192.168.1.10:4321")

# Shorten using a specific Bitly group
groups <- bitly_get_groups()
bitly_shorten("http://192.168.1.10:4321", guid = groups[1])
}

}
\seealso{
\code{\link{bitly_get_groups}}
}
