% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_cache.R
\name{file_cache}
\alias{file_cache}
\title{Create a file cache}
\usage{
file_cache(path, file_id = NULL)
}
\arguments{
\item{path}{Path to a file.}

\item{file_id}{Optional editor file identifier.}
}
\value{
A \code{FileCache} R6 object.
}
\description{
Convenience wrapper for \code{FileCache$new()}.
}
\examples{
\dontrun{
fc <- file_cache("script.R")
fc$content
}
}
\seealso{
\code{\link{FileCache}}
}
