% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/livecode.R
\docType{package}
\name{livecode-package}
\alias{livecode}
\alias{livecode-package}
\title{livecode: Broadcast a source file to multiple concurrant users}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Broadcast a local R (or other text) document over the web and provide live updates as it is edited.
}
\author{
\strong{Maintainer}: Colin Rundel \email{rundel@gmail.com}

}
\keyword{internal}
