% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_ip.R
\name{network_interfaces}
\alias{network_interfaces}
\title{List available network interfaces}
\usage{
network_interfaces()
}
\description{
This function attempts to order the interfaces using a relative priority
based on the ip type, interface type, and then interface name. The goal is
to favor the interface that is most "open" / accessible.
}
