% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_stream_server.R
\name{file_stream_server}
\alias{file_stream_server}
\title{Content-Type' = 'text/html'}
\usage{
file_stream_server(
  host,
  port,
  file,
  file_id,
  interval = 3,
  template = "prism",
  upgrade_content_security_policy
)
}
\description{
Content-Type' = 'text/html'
}
