% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/livecode-package.R, R/livecode.R
\docType{package}
\name{livecode-package}
\alias{livecode}
\alias{livecode-package}
\title{livecode: Broadcast a source file to multiple concurrent users}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

Broadcast a local R (or other text) document over the web and provide live updates as it is edited.

Server for broadcasting source code to multiple viewers
}
\details{
Broadcast a local R (or other text) document over the web and provide live updates as it is edited.
}
\seealso{
\link{serve_file}
}
\author{
\strong{Maintainer}: Colin Rundel \email{rundel@gmail.com}


Colin Rundel \email{rundel@gmail.com}
}
\keyword{internal}
\keyword{package}
