% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/livecode.R
\docType{package}
\name{livecode-package}
\alias{livecode-package}
\alias{livecode}
\title{Source code broadcasting server}
\description{
Server for broadcasting source code to multiple viewers
}
\details{
Broadcast a local R (or other text) document over the web and provide live updates as it is edited.
}
\seealso{
\link{serve_file}
}
\author{
Colin Rundel \email{rundel@gmail.com}
}
\keyword{package}
