% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_stream_server.R
\name{serve_file}
\alias{serve_file}
\title{Create a livecode server for broadcasting a file}
\usage{
serve_file(
  file,
  ip,
  port,
  bitly = FALSE,
  auto_save = TRUE,
  template = "prism",
  interval = 2
)
}
\arguments{
\item{file}{Path to file to broadcast.}

\item{ip}{ip of the server, defaults to the top result of `network_interfaces`.}

\item{port}{port of the server, defaults to a random value.}

\item{bitly}{should a bitly bit link be created for the server.}

\item{auto_save}{should the broadcast file be auto saved update tic.}

\item{template}{page template to use.}

\item{interval}{page update interval in seconds.}
}
\description{
Create a livecode server for broadcasting a file
}
